Remove old explore deep dashboard
