--- app/views/ExploreDeep/ActionCell/styles.css
@@ -1,4 +1,0 @@
-.action-cell {
-    display: flex;
-    justify-content: flex-end;
-}
