--- app/views/ExploreDeep/ActiveProject/styles.css
@@ -1,14 +1,0 @@
-.active-project {
-    border-bottom: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-    padding: var(--dui-spacing-medium);
-
-    .project-title {
-        color: var(--dui-color-brand);
-        font-size: var(--dui-font-size-large);
-        font-weight: var(--dui-font-weight-bold);
-    }
-
-    &:not(:last-child) {
-        border-bottom: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-    }
-}
