--- app/views/ExploreDeep/PublicMapView/ProjectList/styles.css
@@ -1,61 +1,0 @@
-.project-list {
-    display: flex;
-    position: absolute;
-    flex-direction: column;
-    justify-content: center;
-    z-index: 1;
-    border: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-    background-color: var(--dui-color-foreground);
-    width: 100%;
-    max-width: 450px;
-    height: 100%;
-
-    .message {
-        padding: 0;
-    }
-
-    .list {
-        display: flex;
-        flex-direction: column;
-        min-height: 100%;
-        gap: var(--dui-spacing-small);
-
-        .project-container {
-            .heading {
-                color: var(--dui-color-brand);
-            }
-
-            .date {
-                line-height: 2;
-                color: var(--dui-color-text-description);
-                font-weight: var(--dui-font-weight-regular);
-            }
-
-            .content {
-                display: flex;
-                flex-direction: column;
-                gap: var(--dui-spacing-medium);
-            }
-        }
-    }
-
-    .footer {
-        display: flex;
-        align-items: flex-end;
-        flex-direction: column;
-        flex-wrap: wrap;
-        gap: unset;
-
-        .pager {
-            display: flex;
-            justify-content: center;
-            overflow: auto;
-        }
-    }
-
-    .inline {
-        display: flex;
-        justify-content: space-between;
-        margin: var(--dui-spacing-small) 0;
-    }
-}
