--- app/views/ExploreDeep/PublicMapView/styles.css
@@ -1,5 +1,0 @@
-.map-view {
-    display: flex;
-    position: relative;
-    z-index: 0;
-}
