--- app/views/ExploreDeep/PublicTableView/styles.css
@@ -1,4 +1,0 @@
-.table {
-    min-height: 30vh;
-    color: inherit;
-}
