--- app/views/ExploreDeep/TableView/styles.css
@@ -1,4 +1,0 @@
-.table {
-    min-height: 30vh;
-    color: inherit;
-}
