+++ app/views/ExploreDeep/styles.css
@@ -1,131 +1,6 @@
-.explore-deep {
-    padding: var(--dui-spacing-medium) 0 0;
-
-    .header {
-        padding: 0 var(--dui-spacing-super-large);
-    }
-
-    .content {
-        display: flex;
-        flex-direction: column;
-        overflow: unset;
-        gap: var(--dui-spacing-large);
-    }
-
-    .stats-container {
-        display: flex;
-        padding: 0 var(--dui-spacing-super-large);
-        min-height: 40vh;
-        gap: var(--dui-spacing-large);
-
-        .left-container {
-            display: flex;
-            flex-basis: calc(100% / 3 * 2 - var(--dui-spacing-large) / 2);
-            flex-direction: column;
-            gap: var(--dui-spacing-large);
-
-            .info-items-container {
-                display: flex;
-                gap: var(--dui-spacing-large);
-
-                .info-item {
-                    flex-basis: calc(25% - var(--dui-spacing-large) / 3 * 2);
-                }
-
-                .stats-by-time-container {
-                    display: flex;
-                    flex-basis: calc(50% - var(--dui-spacing-large) / 3);
-                    justify-content: center;
-                    gap: var(--dui-spacing-large);
-                    border: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-
-                    .info-item {
-                        flex-basis: calc(100% / 3 - var(--dui-spacing-large));
-                    }
-                }
-            }
-
-            .playback-framework-container {
-                display: flex;
-                flex-grow: 1;
-                gap: var(--dui-spacing-large);
-
-                .playback-card {
-                    display: flex;
-                    align-items: center;
-                    flex-basis: calc(50% - var(--dui-spacing-large) / 2);
-                    justify-content: center;
-                    border: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-                    padding: var(--dui-spacing-large);
-                    gap: var(--dui-spacing-large);
-
-                    .help-message {
-                        max-width: 320px;
-                        line-height: 1.6;
-                    }
-                }
-
-                .framework-container {
-                    display: flex;
-                    flex-basis: calc(50% - var(--dui-spacing-large) / 2);
-                    border: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-                    height: 100%;
-                    overflow-y: auto;
-
-                    .framework-list-container {
-                        padding: 0;
-                        gap: var(--dui-spacing-large);
-                    }
-                }
-            }
-        }
-
-        .header {
-            padding: var(--dui-spacing-medium);
-
-            .heading {
-                gap: var(--dui-spacing-medium);
-            }
-        }
-
-        .right-container {
-            flex-basis: calc(100% / 3 - var(--dui-spacing-large) / 2);
-            border: var(--dui-width-separator-thin) solid var(--dui-color-separator);
-
-            .project-list-container {
-                padding: 0;
-            }
-        }
-    }
-
-    .list {
-        min-height: 100%;
-    }
-
-    .tabs-container {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        background-color: var(--dui-color-foreground);
-        padding: 0 var(--dui-spacing-super-large) var(--dui-spacing-large);
-
-        .filters-container {
-            display: flex;
-            align-items: center;
-            gap: var(--dui-spacing-medium);
-
-            .filters {
-                flex-grow: 1;
-            }
-
-            .tabs {
-                flex-shrink: 0;
-
-                .tab {
-                    padding: var(--dui-spacing-small);
-                    font-size: var(--dui-font-size-extra-large);
-                }
-            }
-        }
-    }
+.export-notification-body {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: var(--dui-spacing-medium);
 }
