--- app/views/NewExploreDeep/styles.css
@@ -1,6 +1,0 @@
-.export-notification-body {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: var(--dui-spacing-medium);
-}
